test: split PDF visual pair ordering cases
